Extract poke workspaces, feedbacks, and template-pull business logic
Move the heavier business-logic blocks out of three poke handlers and
into `lib/api/poke/*`, per [BACK16] (thin handlers) and [BACK17] (shared
between the Next and Hono pair):

- `listWorkspacesForPoke` — search-by-stripe/email/domain/phone +
  plan-priority sort + member-count enrichment. Handler shrinks from
  ~230 lines to ~60 (query parsing + response shaping only).
- `listGlobalAgentFeedbacks` — multi-model fan-in (feedback rows then
  batched workspace / conversation / message lookups). Handler shrinks
  from ~120 lines to ~30.
- `pullTemplatesFromMainRegion` — fetch master-region templates and
  upsert each by handle. Returns `Result<{ count }, "main_region_fetch_failed">`
  so the HTTP shape stays in the handlers.

Types exported from the legacy Next paths (`PokeWorkspaceType`,
`GlobalAgentFeedbackItem`, etc.) are re-exported from `lib/` for
backward compatibility with existing client imports.
